+++ server/vertx/src/test/scala/sttp/tapir/server/vertx/VertxStubServerTest.scala
@@ -4,14 +4,11 @@ import cats.effect.IO
 import cats.effect.std.Dispatcher
 import cats.effect.unsafe.implicits.global
 import sttp.capabilities.fs2.Fs2Streams
-import sttp.capabilities.zio.ZioStreams
 import sttp.client3.testing.SttpBackendStub
 import sttp.monad.FutureMonad
 import sttp.tapir.integ.cats.CatsMonadError
 import sttp.tapir.server.interceptor.CustomiseInterceptors
 import sttp.tapir.server.tests.{CreateServerStubTest, ServerStubStreamingTest, ServerStubTest}
-import zio.stream.ZStream
-import zio.{Runtime, Task}
 
 import scala.concurrent.{ExecutionContext, Future}
 
@@ -41,18 +38,4 @@ class VertxCatsServerStubStreamingTest extends ServerStubStreamingTest(new Vertx
 
   /** Must be an instance of streams.BinaryStream */
   override def sampleStream: Any = fs2.Stream.apply[IO, String]("hello")
-}
-
-object VertxZioCreateServerStubTest extends CreateServerStubTest[Task, VertxZioServerOptions[Task]] {
-  override def customiseInterceptors: CustomiseInterceptors[Task, VertxZioServerOptions[Task]] = VertxZioServerOptions.customiseInterceptors
-  override def stub[R]: SttpBackendStub[Task, R] = SttpBackendStub(VertxZioServerInterpreter.monadError)
-  override def asFuture[A]: Task[A] => Future[A] = task => Runtime.default.unsafeRunToFuture(task)
-}
-
-class VertxZioServerStubTest extends ServerStubTest(VertxZioCreateServerStubTest)
-
-class VertxZioServerStubStreamingTest extends ServerStubStreamingTest(VertxZioCreateServerStubTest, ZioStreams) {
-
-  /** Must be an instance of streams.BinaryStream */
-  override def sampleStream: Any = ZStream.fromIterable(List("hello"))
 }
